Removed Spring usage in calculations tests refactor
